tools: Fix Coverity errors in resolution plugin.
Fix multiplication overflow and fgetc usage to avoid Coverity errors.

+++ open-vm-tools/services/plugins/resolutionSet/resolutionX11.c
@@ -533,7 +533,7 @@ SelectResolution(uint32 width,
     * and height.
     */
    for (i = 0; i < xrrNumSizes; i++) {
-      potentialSize = xrrSizes[i].width * xrrSizes[i].height;
+      potentialSize = (uint64)xrrSizes[i].width * xrrSizes[i].height;
       if (xrrSizes[i].width <= width && xrrSizes[i].height <= height &&
           potentialSize > bestFitSize ) {
          bestFitSize = potentialSize;
